Run Chronyd with restart monitor
